Add href to asana anchor workaround (#155)
# Add href to asana anchor workaround

## Summary
The asana client cannot properly handle links to deleted objects that
directly reference the global id of the object. This manifests in errors
that are difficult to handle when PR bodies contain references to
de-activated users. Most frequently, this results is a slew of SGTM
errors after intern season.

Instead, we can create an actual link to the github handle with a
`data-asana-gid` attribute. If the user_id is valid, the anchor tag
should not _do_ anything and be replaced with the appropriate asana task
link. If the user_id is _not_ valid, then we should be able to use the
anchor tag's href as a valid link.

## Test plan
- [x] a PR that references a de-activated asana user will properly link
the github handle link and succeed in upserting the task
- [x] a PR that references an activated user still references the asana
task and not the github handle

<img width="636" alt="image"
src="https://github.com/Asana/SGTM/assets/22352112/d767b49b-fb52-4ff4-ade2-514ac4890150">
https://app.asana.com/0/1127984365697766/1205259139814521


Pull Request synchronized with [Asana
task](https://app.asana.com/0/0/1205380760704777)