Refactor graphql queries (#157)
This is a prerequisite PR for a combined refactor where we no longer
read the statusCheckRollup state and instead rollup our own state.

statusCheckRollup status ignoring requires additional information in
order to not rerun redundant check runs. We want to eventually replace
the base ref with `isRequired` queries since those can be configured via
github instead of through SGTM.

I've separated this out as mostly a refactor except changing the
following functionality to be less redundant:

`get_pull_request_for_commit` now iterates through all the associated
pull requests and matches the pull requests with their last commit being
the same commit Id we're querying with. This helps to reduce redundancy
since we always want to get the last "event" on the pull request.

If an author has a pull request that contains the commit and another
pull request that has the commit as head, SGTM would previously only
handle the older pull request. Instead, we want to handle the first pull
request where the head is the same commit to capture the most recent
event on the pull request.

## Test Plan
added new client tests for the modification to
`get_pull_request_for_commit`


Pull Request synchronized with [Asana
task](https://app.asana.com/0/0/1205687342254678)